moving the registration of commands out of plugin and into the class
