+++ tythan-bukkit/src/main/java/co/lotc/core/bukkit/TythanBukkit.java
@@ -7,23 +7,16 @@ import org.bukkit.plugin.java.JavaPlugin;
 
 import co.lotc.core.Tythan;
 import co.lotc.core.TythanProvider;
-import co.lotc.core.agnostic.Command;
-import co.lotc.core.bukkit.command.ArcheCommandExecutor;
 import co.lotc.core.bukkit.command.BrigadierProvider;
-import co.lotc.core.bukkit.command.BukkitKommandant;
 import co.lotc.core.bukkit.command.ItemArg;
 import co.lotc.core.bukkit.command.SenderTypes;
 import co.lotc.core.bukkit.menu.MenuListener;
 import co.lotc.core.bukkit.util.ChatBuilder;
 import co.lotc.core.bukkit.util.Run;
 import co.lotc.core.bukkit.util.WeakBlock;
-import co.lotc.core.bukkit.wrapper.BukkitCommand;
 import co.lotc.core.bukkit.wrapper.BukkitConfig;
-import co.lotc.core.command.ArcheCommand;
 import co.lotc.core.command.brigadier.CommandNodeManager;
-import co.lotc.core.command.brigadier.Kommandant;
 import lombok.Getter;
-import lombok.var;
 
 public class TythanBukkit extends JavaPlugin implements Tythan {
 	public static TythanBukkit get(){
@@ -75,18 +68,7 @@ public class TythanBukkit extends JavaPlugin implements Tythan {
 	public ChatBuilder chatBuilder() {
 		return new ChatBuilder();
 	}
-	
-	@Override
-	public void registerRootCommand(Command wrapper, ArcheCommand handler) {
-		Kommandant kommandant = new BukkitKommandant(handler);
-		kommandant.addBrigadier();
-		CommandNodeManager.getInstance().register(kommandant);
-		
-		var pluginCommand = ((BukkitCommand) wrapper).getHandle();
-		ArcheCommandExecutor executor = new ArcheCommandExecutor(handler);
-		pluginCommand.setExecutor(executor);
-	}
-	
+
 	private void listen(Listener l) {
 		Bukkit.getPluginManager().registerEvents(l, this);
 	}
