+++ tythan-common/src/main/java/co/lotc/core/Tythan.java
@@ -4,9 +4,7 @@ import java.io.File;
 import java.util.logging.Logger;
 
 import co.lotc.core.agnostic.AbstractChatBuilder;
-import co.lotc.core.agnostic.Command;
 import co.lotc.core.agnostic.Config;
-import co.lotc.core.command.ArcheCommand;
 
 public interface Tythan {
 	static Tythan get() { return TythanProvider.INSTANCE; }
@@ -20,6 +18,4 @@ public interface Tythan {
 	Config config();
 	
 	AbstractChatBuilder<?> chatBuilder();
-	
-	void registerRootCommand(Command wrapper, ArcheCommand handler);
 }
